feat: Implement AWS FinOps Dashboard with audit and trend analysis features
- Added dashboard_runner.py to handle the main dashboard logic, including profile initialization, audit report generation, and trend analysis.
- Introduced profile_processor.py for processing individual and combined AWS profiles, fetching cost data, and summarizing EC2 instances.
- Created configuration files (config.json, config.toml, config.yaml, config.yml) to allow users to customize profiles, regions, report types, and other settings.
- Enhanced error handling and logging for better user feedback during profile processing and data fetching.
- Integrated cost data visualization with rich console output for improved user experience.
