+++ aws_finops_dashboard/types.py
@@ -27,7 +27,7 @@ class CostData(TypedDict):
     current_period_end: str
     previous_period_start: str
     previous_period_end: str
-    monthly_costs: List[Tuple[str, float]]
+    monthly_costs: Optional[List[Tuple[str, float]]]
 
 
 class ProfileData(TypedDict):
@@ -46,6 +46,7 @@ class ProfileData(TypedDict):
     error: Optional[str]
     current_period_name: str
     previous_period_name: str
+    percent_change_in_total_cost: Optional[float]
 
 
 class CLIArgs(TypedDict, total=False):
