+++ config.yaml
@@ -1,0 +1,16 @@
+# Example config.yaml for AWS FinOps Dashboard
+
+profiles:
+  - default
+regions:
+  - us-east-1
+  - eu-west-2
+combine: true # Combine profiles from the same AWS account
+report_name: "monthly_finops_summary"
+report_type:
+  - csv
+  - pdf
+dir: "./reports/aws-finops"
+time_range: 30 # Time range for cost data in days
+audit: true # Run the audit report
+trend: false # Do not run the trend report by default
