+++ requirements.txt
@@ -12,4 +12,3 @@ urllib3==2.3.0
 requests>=2.0.0
 packaging>=21.0
 reportlab>=3.6.1
-
